+++ app/src/main/java/edu/sc/FitLivin/StrengthDayTwo.java
@@ -166,10 +166,10 @@ MediaPlayer mp;
                                             public void done(List<ParseObject> objects, ParseException e) {
                                                 if (e == null && objects.size() != 0) {
                                                     if (objects.get(0).get("username").equals(ParseUser.getCurrentUser().getUsername())) {
-                                                        int x = (Integer) objects.get(0).get("CurrentPoints");
+                                                        Double x = (Double) objects.get(0).get("CurrentPoints");
                                                         MainActivity main = new MainActivity();
                                                         main.points = x;
-                                                        Integer points = main.points;
+                                                        Double points = main.points;
                                                         points = points + 50;
                                                         String s = ParseUser.getCurrentUser().getUsername();
                                                         main.pointsData(points,s);
